requirements.txt
argparse
numpy
pandas
matplotlib
torch
torchvision
torchaudio
pillow